Generate mobi and epub next to pdf as per #28.
Note that this does _not yet_ produce a 1:1 version of what you can see
in the PDF: Chapter authors and biographies are currently being swallowed
due to heavily customized tags. This will be addressed in a separate
commit, Also, meta data are missing.

Compared to the book, the mobi and epub versions lack the cover page.